Implemented a new way to handle errors.
Errors are not longer as Either; also, non existing resources are not
Optional anymore. Everything will be used as a MonadError. In this way
a unexisting Protocol will raise an error `Sync[F].raiseError` and the
computation will finish with that error. For handling errors, I have
implemented a middleware at route level so what it does is to capture
the computation error result and convert it into the adequate http
response.

This is a simple and friendly version of exposed here:
https://typelevel.org/blog/2018/11/28/http4s-error-handling-mtl-2.html